Reorganice y corre!

+++ Kernel/util/include/util.h
@@ -5,7 +5,7 @@ char *itoa(int value, char *str, int base);
 int intlen(int n);
 int strlen(char *s);
 
-int to_bcd8(int);
-int to_bcd16(int);
+uint8_t to_bcd8(uint8_t);
+uint16_t to_bcd16(uint16_t);
 
 #endif
